+++ src/pages/About.jsx
@@ -38,10 +38,10 @@ const About = () => {
               <h3 className="highlighter">
                 <div>Frontend developer currently based in Toronto, Canada. Mostly focused on working on the front-end side of things including UI/UX and graphic design.</div>
               </h3>
-              <p>I began my front-end journey in the early 2000s as a high school student. My initial thought was that I wanted to put together a website for my class as we just got our first Internet connection. I used the gool ol' <b>Microsoft FrontPage Express</b> that came with Windows and signed up for a free <b>GeoCities</b> (!) account.</p>
-              <p>It was that project that started off as a hobby and it eventually ignited my interest in web development, and the rest was history. I've been a front-end developer for the last ten years. I've specialized in UI/UX and client interaction while working with frameworks and languages including React, HTML5, CSS, SASS, PHP, Laravel and WordPress.</p>
-              <p>I've worked closely with numerous design and consulting organizations over the past decade, overseeing and contributing to projects in a variety of roles. These included application management, digital transformations, and redesigns of company websites. I'm particularly passionate about modernizing legacy systems and streamlining business processes.</p>
-              <p>My work on the visual side of things has fueled my interest in multimedia and engaging user experiences. I'm currently expanding my skillset in UI/UX to blend creativity with logic and turning the subjective into objective.</p>
+              <p className='generic'>I began my front-end journey in the early 2000s as a high school student. My initial thought was that I wanted to put together a website for my class as we just got our first Internet connection. I used the gool ol' <b>Microsoft FrontPage Express</b> that came with Windows and signed up for a free <b>GeoCities</b> (!) account.</p>
+              <p className='generic'>It was that project that started off as a hobby and it eventually ignited my interest in web development, and the rest was history. I've been a front-end developer for the last ten years. I've specialized in UI/UX and client interaction while working with frameworks and languages including React, HTML5, CSS, SASS, PHP, Laravel and WordPress.</p>
+              <p className='generic'>I've worked closely with numerous design and consulting organizations over the past decade, overseeing and contributing to projects in a variety of roles. These included application management, digital transformations, and redesigns of company websites. I'm particularly passionate about modernizing legacy systems and streamlining business processes.</p>
+              <p className='generic'>My work on the visual side of things has fueled my interest in multimedia and engaging user experiences. I'm currently expanding my skillset in UI/UX to blend creativity with logic and turning the subjective into objective.</p>
               <div className="actions">
                 <Link className="g-action" target="_blank" to="#"><i className="fa-solid fa-file-pdf"></i> My Resume <i className="fa-solid fa-chevron-right"></i></Link>
               </div>
@@ -159,24 +159,24 @@ const About = () => {
               <div className="col-a">
                 { (lang === "EN") ? (
                   <>
-                    <p>When I'm not busy crafting digital experiences, you'll often find me enjoying movies and video games on Twitch with friends, doing casual cosplays at comic conventions, or just walking around exploring Toronto while listening to J-Pop and EDM. I also make time for running, (casual) photography, and exploring new places! My travels have taken me to the United States, Japan, Hong Kong, and Macau so far&mdash; and this is just the beginning!</p>
-                    <p>I'm currently learning Japanese, so you might spot the occasional Japanese post (with translations, of course!).</p>
-                    <p>While I'm not [yet] an avid reader, I enjoy a good read. Some of my favourite authors include Kyle Higgins, Dan Brown and Suzanne Collins. I recently finished reading <i>Mary and the Secret Garden</i> and I'm currently hooked on Boom! series' <i>Power Rangers Comics</i> universe.</p>
+                    <p className='generic'>When I'm not busy crafting digital experiences, you'll often find me enjoying movies and video games on Twitch with friends, doing casual cosplays at comic conventions, or just walking around exploring Toronto while listening to J-Pop and EDM. I also make time for running, (casual) photography, and exploring new places! My travels have taken me to the United States, Japan, Hong Kong, and Macau so far&mdash; and this is just the beginning!</p>
+                    <p className='generic'>I'm currently learning Japanese, so you might spot the occasional Japanese post (with translations, of course!).</p>
+                    <p className='generic'>While I'm not [yet] an avid reader, I enjoy a good read. Some of my favourite authors include Kyle Higgins, Dan Brown and Suzanne Collins. I recently finished reading <i>Mary and the Secret Garden</i> and I'm currently hooked on Boom! series' <i>Power Rangers Comics</i> universe.</p>
                   </>
                 ) : (<></>) }
                 { (lang === "JP") ? (
                   <>
-                    <small className="jp"><b>警告:</b> 文法が間違っている可能性があります。</small>
-                    <p className="jp">パソコンの前に忙しくないときは、友達と一緒に映画を見たり、Twitchでビデオゲームを楽しんだり、コミコンでコスプレをしたり、J-POPやEDMなどを聞きながらトロントを散歩してる事が多いです。また、時々ランニングや写真撮影、いろいろな新しい所を観光にも時間を割いています。今までアメリカ、日本、香港、マカオにいたことがあるんだけど、これはまだ冒険の始まりに過ぎません!</p>
-                    <p className="jp">現在日本語を勉強しているので、このサイトで時々日本語で練習すると日本語の投稿が見られるかもしれません（訳するつもり、大丈夫なんです!）。</p>
-                    <p className="jp">まだ本の虫にならないんですが、いい本を読むのが好きです。お気に入りの作家にはカイル・ヒギンズ、ダン・ブラウン、スザンヌ・コリンズなどがいます。最近「秘密の花園」を読み終え、現在はブーンシリーズの「パワーレンジャー」の漫画を読み込んでいます。</p>
+                    <small className="generic jp"><b>警告:</b> 文法が間違っている可能性があります。</small>
+                    <p className="generic jp">パソコンの前に忙しくないときは、友達と一緒に映画を見たり、Twitchでビデオゲームを楽しんだり、コミコンでコスプレをしたり、J-POPやEDMなどを聞きながらトロントを散歩してる事が多いです。また、時々ランニングや写真撮影、いろいろな新しい所を観光にも時間を割いています。今までアメリカ、日本、香港、マカオにいたことがあるんだけど、これはまだ冒険の始まりに過ぎません!</p>
+                    <p className="generic jp">現在日本語を勉強しているので、このサイトで時々日本語で練習すると日本語の投稿が見られるかもしれません（訳するつもり、大丈夫なんです!）。</p>
+                    <p className="generic jp">まだ本の虫にならないんですが、いい本を読むのが好きです。お気に入りの作家にはカイル・ヒギンズ、ダン・ブラウン、スザンヌ・コリンズなどがいます。最近「秘密の花園」を読み終え、現在はブーンシリーズの「パワーレンジャー」の漫画を読み込んでいます。</p>
                   </>
                 ) : (<></>) }
                 { (lang === "PH") ? (
                   <>
-                    <p>Kapag wala ako sa harap ng computer, madalas akong lumalabas at nanood ng pelikula o naglalaro ng video games sa Twitch kasama ang mga kaibigan, nagko-cosplay sa mga conventions or kaya naglalakad lang sa Toronto habang nakikinig ng J-Pop at EDM. Minsan tumatakbo din ako, kumukuha ng litrato at namamasyal sa iba't-ibang lugar. Nakapasyal na ako sa Amerika, bansang Hapon, Hong Kong at Macau&mdash; at madami pa akong gustong mapasyalan!</p>
-                    <p>Kasalukuyan akong nag-aaral ng wikang Hapon, kaya minsan ay mapapansin ninyong nagpa-practice ako dito (may translation naman, okey lang yan!)</p>
-                    <p>Hindi pa naman ako masyadong nahuhumaling sa pagbabasa, pero may mga nagugustuhan naman ako. Paborito kong basahin yung mga gawa nila Kyle Higgins, Dan Brown at Suzanne Collins. Kakatapos ko lang basahin yung <i>Mary and the Secret Garden</i> at ngayon binabasa ko yung komics ng <i>Power Rangers</i> na gawa ng Boom! Studios.</p>
+                    <p className='generic'>Kapag wala ako sa harap ng computer, madalas akong lumalabas at nanood ng pelikula o naglalaro ng video games sa Twitch kasama ang mga kaibigan, nagko-cosplay sa mga conventions or kaya naglalakad lang sa Toronto habang nakikinig ng J-Pop at EDM. Minsan tumatakbo din ako, kumukuha ng litrato at namamasyal sa iba't-ibang lugar. Nakapasyal na ako sa Amerika, bansang Hapon, Hong Kong at Macau&mdash; at madami pa akong gustong mapasyalan!</p>
+                    <p className='generic'>Kasalukuyan akong nag-aaral ng wikang Hapon, kaya minsan ay mapapansin ninyong nagpa-practice ako dito (may translation naman, okey lang yan!)</p>
+                    <p className='generic'>Hindi pa naman ako masyadong nahuhumaling sa pagbabasa, pero may mga nagugustuhan naman ako. Paborito kong basahin yung mga gawa nila Kyle Higgins, Dan Brown at Suzanne Collins. Kakatapos ko lang basahin yung <i>Mary and the Secret Garden</i> at ngayon binabasa ko yung komics ng <i>Power Rangers</i> na gawa ng Boom! Studios.</p>
                   </>
                 ) : (<></>) }
               </div>