+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
-## [0.2.5] - 2024-08-30
+## [0.2.5] - 2024-09-01
 ### Added
 - [About] PH language support
 - [Twitch] Embedded player
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [Twitch] Now Streaming + Community
 ### Changed
 - [Global] Cleaned up global styles
+- [Global] Fixed pages to top scroll for each load
+- [About] Updated/cleaned layout
 
 ## [0.2.4] - 2024-08-28
 ### Added